requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.12.0
cryptography>=42.0.0
pytest>=8.0.0